Move constants files to components folder
