+++ src/components/Skills/Skills.jsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect, useRef } from 'react';
 import { motion, AnimatePresence } from 'framer-motion';
-import { web, game, design, skill } from '../Data/skillsData';
+import { web, game, design, skill } from './Skills.constants';
 import Counter from '../Counter/Counter';
 import SkillItem from './SkillItem/SkillItem';
 import { skillsAnim, leftAnim, rightAnim, skillCatAnim } from './Skills.anim';
